Walk back file separation to simplify this PR
