+++ man/vec_split.Rd
@@ -1,5 +1,5 @@
 % Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/split.R
+% Please edit documentation in R/dictionary.R
 \name{vec_split}
 \alias{vec_split}
 \title{Split a vector into groups}
