+++ man/vec_split_id.Rd
@@ -1,5 +1,5 @@
 % Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/split.R
+% Please edit documentation in R/dictionary.R
 \name{vec_split_id}
 \alias{vec_split_id}
 \title{Locate unique groups in a vector}
