+++ src/vctrs.h
@@ -211,10 +211,7 @@ R_len_t vec_bare_dim_n(SEXP x);
 SEXP vec_cast(SEXP x, SEXP to, struct vctrs_arg* x_arg, struct vctrs_arg* to_arg);
 SEXP vec_cast_common(SEXP xs, SEXP to);
 SEXP vec_coercible_cast(SEXP x, SEXP to, struct vctrs_arg* x_arg, struct vctrs_arg* to_arg);
-SEXP vec_as_index(SEXP i, R_len_t n, SEXP names);
 SEXP vec_slice(SEXP x, SEXP index);
-SEXP vec_slice_fallback(SEXP x, SEXP index);
-SEXP vec_slice_base(enum vctrs_type type, SEXP x, SEXP index);
 SEXP vec_split_along(SEXP x, SEXP indices);
 SEXP vec_slice_shaped(enum vctrs_type type, SEXP x, SEXP index);
 SEXP vec_assign(SEXP x, SEXP index, SEXP value);
@@ -231,9 +228,6 @@ SEXP vec_type2(SEXP x,
                struct vctrs_arg* x_arg,
                struct vctrs_arg* y_arg,
                int* left);
-
-SEXP slice_names(SEXP names, SEXP index);
-SEXP slice_rownames(SEXP names, SEXP index);
 
 bool is_data_frame(SEXP x);
 bool is_record(SEXP x);
