+++ tests/testthat/test-dictionary.R
@@ -191,3 +191,104 @@ test_that("matching functions work with different encodings", {
   expect_equal(vec_match(encs, encs[1]), rep(1, 3))
   expect_equal(vec_in(encs, encs[1]), rep(TRUE, 3))
 })
+
+# splits ------------------------------------------------------------------
+
+test_that("can split empty vector", {
+  out <- vec_split(integer(), character())
+
+  expect_s3_class(out, "data.frame")
+  expect_equal(out$key, character())
+  expect_equal(out$val, list_of(.ptype = integer()))
+})
+
+test_that("split data frame with data frame", {
+  df <- data.frame(x = c(1, 1, 2), y = c(1, 1, 1))
+  out <- vec_split(df, df)
+
+  expect_s3_class(out, "data.frame")
+  expect_equal(out$key, data.frame(x = c(1, 2), y = c(1, 1)))
+  expect_equal(out$val, list_of(
+    data.frame(x = c(1, 1), y = c(1, 1)),
+    data.frame(x = 2, y = 1)
+  ))
+})
+
+test_that("x and by must be same size", {
+  expect_error(
+    vec_split(1:3, 1:2),
+    "same size"
+  )
+})
+
+test_that("split takes the equality proxy (#375)", {
+  scoped_comparable_tuple()
+  x <- tuple(c(1, 2, 1), 1:3)
+  expect_identical(nrow(vec_split(1:3, x)), 2L)
+})
+
+test_that("split works with different encodings", {
+  encs <- encodings()
+  expect_identical(nrow(vec_split(1:3, encs)), 1L)
+})
+
+# split id ---------------------------------------------------------------
+
+test_that("can locate unique groups of an empty vector", {
+  out <- vec_split_id(integer())
+
+  expect_s3_class(out, "data.frame")
+  expect_equal(out$key, integer())
+  expect_equal(out$id, list_of(.ptype = integer()))
+})
+
+test_that("can locate unique groups of a data frame", {
+  df <- data_frame(x = c(1, 1, 1, 2, 2), y = c("a", "a", "b", "a", "b"))
+  out <- vec_split_id(df)
+
+  expect_equal(nrow(out), 4L)
+  expect_equal(out$key, vec_unique(df))
+})
+
+test_that("can locate unique groups of a data frame with a list column", {
+  df <- data_frame(x = list(1:2, 1:2, "a", 5.5, "a"))
+  out <- vec_split_id(df)
+
+  expect_equal(nrow(out), 3L)
+  expect_equal(out$key, vec_unique(df))
+})
+
+test_that("`x` must be a vector", {
+  expect_error(vec_split_id(environment()), class = "vctrs_error_scalar_type")
+})
+
+test_that("`key` column retains full type information", {
+  x <- factor(letters[c(1, 2, 1)], levels = letters[1:3])
+  out <- vec_split_id(x)
+
+  expect_equal(levels(out$key), levels(x))
+})
+
+test_that("`key` and `value` retain names", {
+  x <- c(a = 1, b = 2, c = 1, a = 1)
+  split <- vec_split(x, x)
+  expect_identical(split$key, c(a = 1, b = 2))
+  expect_identical(split$val[[1]], c(a = 1, c = 1, a = 1))
+  expect_identical(split$val[[2]], c(b = 2))
+})
+
+test_that("vec_split_id takes the equality proxy", {
+  scoped_comparable_tuple()
+  x <- tuple(c(1, 2, 1), 1:3)
+  expect_equal(vec_split_id(x)$key, x[1:2])
+  expect_equal(vec_split_id(x)$id, list_of(c(1L, 3L), 2L))
+
+  x <- as.POSIXlt(new_datetime(c(1, 2, 1)))
+  expect_equal(vec_split_id(x)$key, x[1:2])
+  expect_equal(vec_split_id(x)$id, list_of(c(1L, 3L), 2L))
+})
+
+test_that("vec_split_id works with different encodings", {
+  encs <- encodings()
+  expect_identical(nrow(vec_split_id(encs)), 1L)
+})
